requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==20.1.0
requests==2.31.0
google-auth==2.21.0
google-api-python-client==2.79.0
python-dotenv==1.0.0

